requirements.txt
# ============================================================
# Maju Jaya Data Platform — Core Dependencies
# Python >= 3.11
# Airflow BUKAN di sini — jalan via Docker (docker-compose.yml)
# Install: uv pip install -r requirements.txt
# ============================================================

# ── DATA PROCESSING ──────────────────────────────────────────
pandas==2.1.4
pyarrow==17.0.0
numpy==1.26.4
openpyxl==3.1.5            # Baca Excel (.xlsx)

# ── MYSQL SOURCE ─────────────────────────────────────────────
pymysql==1.1.0
sqlalchemy==2.0.36

# ── GOOGLE CLOUD ─────────────────────────────────────────────
google-cloud-storage==2.18.2
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.27.0
google-api-python-client==2.114.0   # Google Drive API
google-auth-oauthlib==1.2.0         # OAuth untuk Drive
db-dtypes==1.3.1
google-auth==2.36.0

# ── DBT ──────────────────────────────────────────────────────
dbt-core==1.8.9
dbt-bigquery==1.8.3

# ── GREAT EXPECTATIONS ───────────────────────────────────────
great-expectations==0.18.21

# ── TESTING ──────────────────────────────────────────────────
pytest==8.3.4

# ── DEV UTILITIES ────────────────────────────────────────────
python-dotenv==1.0.1
rich==13.9.4